Bugfix: gluecode model options (#113)
* hotfix

* add the actual hotfix

* update all examples

* added example testing routine to test

* adjust cwd command

* reorganize some examples

* attempt fix

* black reformat, pip statsmodels again after their hotfix

* ok think i fixed it

* cleanup runner

* try system-agnostic runner

* try system-agnostic runner again

* indent fix

* turn off n2 diagrams

* test consolidated CI/CD workflow

* attempt improved consolidated script

* direct parallel example runner test

* complete runner code

* a new test that should maybe actually work

* add missing file and quotation marks

* switch to consolidated with parallel examples

* rectify author attribution in pyproject.toml

* remove commented dependency

* cleaner gui interaction

* modified consolidated tests

* faster consolidated ci/cd

* rollback action

* take 2 on caching

* update dependency

* slight bugfix mod

* try cache again

* iterating on cicd cache

* reorder

* rename
 test

* address some of jareds comments and remove commented code

* black reformat
