--- .github/workflows/python-tests-system.yml
@@ -1,47 +1,0 @@
-name: Python System Tests with Coverage
-
-on:
-  push:
-    branches:
-      - '*'
-  pull_request:
-    branches:
-      - '*'
-
-jobs:
-  test:
-    strategy:
-      matrix:
-        python-version: [3.12] # ["3.10", "3.11", "3.12", "3.13"]
-        os: [macos-latest, ubuntu-latest, windows-latest]
-    runs-on: ${{ matrix.os }}
-
-    steps:
-    - name: Checkout code
-      uses: actions/checkout@v4
-
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v5
-      with:
-        python-version: ${{ matrix.python-version }}
-
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-
-    - name: Install Ard
-      run: |
-        pip install .[dev]
-
-    - name: Run system tests with coverage
-      run: |
-        pytest --cov=ard --cov-fail-under=50 test/system
-        # pytest --cov=ard --cov-fail-under=80 test/system
-
-    # - name: Upload coverage report
-    #   uses: codecov/codecov-action@v2
-    #   with:
-    #     file: ./coverage.xml
-    #     flags: unittests
-    #     name: codecov-umbrella
-
