+++ ard/__init__.py
@@ -6,5 +6,5 @@ from . import utils
 
 from pathlib import Path
 
-BASE_DIR = Path(__file__).resolve().parent
+BASE_DIR = Path(__file__).absolute().parent
 ASSET_DIR = BASE_DIR / "api" / "default_systems"
