--- examples/01-iea-wind-740-10-rowp/inputs/ard/ard_system.yaml
@@ -1,16 +1,0 @@
-groups:
-  aerodynamics:
-    promotes: ["*"]
-    systems:
-      aep:
-        type: group
-        promotes: ["*"]
-        systems:
-          floris:
-            name: farm_aero.floris
-            object: FLORISAEP
-            type: component
-            promotes: ["*"]
-            kwargs:
-              case_title: "test"
-              modeling_options: "example_value"
