--- examples/01-iea-wind-740-10-rowp/inputs/data/IEA37_10MW_turbine.yaml
@@ -1,14 +1,0 @@
-name: IEA Wind Task 37 10MW Offshore Reference Turbine
-performance:
-  rated_power: 10000000
-  cutin_wind_speed: 4.0
-  cutout_wind_speed: 25.0
-  power_curve: 
-    power_values: [387510.9723, 644692.8925, 937974.799, 1257928.478, 1604119.995, 1969504.269, 2340111.432, 2708595.63, 3067098.715, 3408654.893, 3723706.607, 4003354.317, 4239645.746, 4425830.055, 4556570.854, 4628112.983, 4638399.089, 4690045.495, 4804899.691, 4985748.651, 5236938.939, 5564425.114, 5975831.243, 6480522.086, 7089679.768, 7816381.143, 8690612.607, 9716655.499, 10000000, 9999999.662, 10000000.11, 9999986.888, 10000034.43, 10000023.5, 10000017.75, 10000014.23, 10000011.69, 10000009.65, 10000007.96, 10000006.47, 10000005.1, 10000003.98, 10000003.09, 10000002.36, 10000001.76, 10000001.26, 10000000.85, 10000000.53, 10000000.28, 10000041.66]
-    power_wind_speeds: [4, 4.514652562, 5.000795688, 5.457350016, 5.883301884, 6.277705571, 6.639685402, 6.968437691, 7.263232527, 7.523415391, 7.748408611, 7.937712648, 8.090907197, 8.20765213, 8.287688244, 8.330837837, 8.337005108, 8.367833853, 8.435590166, 8.540123611, 8.681202099, 8.858512398, 9.071660837, 9.320174171, 9.603500641, 9.921011189, 10.27200086, 10.65569038, 10.75773635, 11.51769068, 11.9940876, 12.49936091, 13.03238875, 13.59198769, 14.17691527, 14.7858728, 15.41750825, 16.07041923, 16.74315611, 17.43422525, 18.1420923, 18.86518562, 19.60189976, 20.35059904, 21.10962115, 21.87728087, 22.65187381, 23.43168018, 24.21496861, 25]
-  Ct_curve:
-    Ct_values: [0.770113776, 0.776301765, 0.782430404, 0.781982993, 0.780246928, 0.777153006, 0.771853687, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.776845963, 0.767521911, 0.765104347, 0.758675026, 0.505649564, 0.431035612, 0.370827539, 0.320855929, 0.278805531, 0.243151019, 0.212773558, 0.186803785, 0.164542221, 0.145414565, 0.128943645, 0.114730283, 0.102438991, 0.091786964, 0.082535328, 0.074482009, 0.067455867, 0.061311766, 0.05592648, 0.051195261, 0.047029125]
-    Ct_wind_speeds: [4, 4.514652562, 5.000795688, 5.457350016, 5.883301884, 6.277705571, 6.639685402, 6.968437691, 7.263232527, 7.523415391, 7.748408611, 7.937712648, 8.090907197, 8.20765213, 8.287688244, 8.330837837, 8.337005108, 8.367833853, 8.435590166, 8.540123611, 8.681202099, 8.858512398, 9.071660837, 9.320174171, 9.603500641, 9.921011189, 10.27200086, 10.65569038, 10.75773635, 11.51769068, 11.9940876, 12.49936091, 13.03238875, 13.59198769, 14.17691527, 14.7858728, 15.41750825, 16.07041923, 16.74315611, 17.43422525, 18.1420923, 18.86518562, 19.60189976, 20.35059904, 21.10962115, 21.87728087, 22.65187381, 23.43168018, 24.21496861, 25]
-hub_height: 119.0
-rotor_diameter: 198.0
-tilt: 5.0
