--- examples/01-iea-wind-740-10-rowp/inputs/data/ROWP_Regular.yaml
@@ -1,64 +1,0 @@
-name: IEA Wind 740-10-MW Reference Offshore Wind Plant with regular layout (plant)
-layouts:
-    initial_layout:
-        coordinates:
-            x: [
-                500968.1461, 499748.6565, 501245.7744, 500026.2848, 498527.8286,
-                497308.339, 501523.4027, 500303.9131, 498805.4569, 497585.9673,
-                496087.5111, 494868.0215, 501801.031, 500581.5414, 499083.0852,
-                497863.5956, 496365.1394, 495145.6498, 493647.1936, 492427.704,
-                502078.6593, 500859.1697, 499360.7135, 498141.2239, 496642.7677,
-                495423.2781, 493924.8219, 492705.3323, 491206.8762, 489987.3865,
-                502356.2876, 501136.798, 499638.3418, 498418.8522, 496920.396,
-                495700.9064, 494202.4502, 492982.9606, 491484.5045, 490265.0148,
-                488766.5587, 487547.069, 502633.9159, 501414.4263, 499915.9701,
-                498696.4805, 497198.0243, 495978.5347, 494480.0786, 493260.5889,
-                491762.1328, 490542.6431, 489044.187, 487824.6973, 486326.2412,
-                485106.7515, 497475.6526, 496256.163, 494757.7069, 493538.2172,
-                492039.7611, 490820.2714, 489321.8153, 488102.3256, 486603.8695,
-                497753.2809, 496533.7913, 495035.3352, 493815.8455, 492317.3894,
-                489599.4436, 498030.9093, 496811.4196, 495312.9635
-            ]
-            y: [
-                5716452.784, 5717635.848, 5718427.22, 5719610.283, 5718809.394,
-                5719992.458, 5720401.656, 5721584.719, 5720783.83, 5721966.894,
-                5721166.004, 5722349.068, 5722376.092, 5723559.155, 5722758.266,
-                5723941.33, 5723140.44, 5724323.504, 5723522.615, 5724705.678,
-                5724350.528, 5725533.591, 5724732.702, 5725915.765, 5725114.876,
-                5726297.94, 5725497.051, 5726680.114, 5725879.225, 5727062.288,
-                5726324.963, 5727508.027, 5726707.138, 5727890.201, 5727089.312,
-                5728272.376, 5727471.486, 5728654.55, 5727853.661, 5729036.724,
-                5728235.835, 5729418.899, 5728299.399, 5729482.463, 5728681.574,
-                5729864.637, 5729063.748, 5730246.812, 5729445.922, 5730628.986,
-                5729828.097, 5731011.16, 5730210.271, 5731393.335, 5730592.445,
-                5731775.509, 5731038.184, 5732221.248, 5731420.358, 5732603.422,
-                5731802.533, 5732985.596, 5732184.707, 5733367.77, 5732566.881,
-                5733012.62, 5734195.683, 5733394.794, 5734577.858, 5733776.968,
-                5734159.143, 5734987.056, 5736170.119, 5735369.23
-            ]
-                
-electrical_substations:
-    coordinates:
-        x: [497620.7]
-        y: [5730622.0]
-electrical_collection_array:
-    edges: [[0, 2, 0], [1, 4, 0], [2, 6, 0], [3, 8, 0], [4, 5, 0], [5, 10, 0], [6, 7, 0],
-            [7, 14, 1], [8, 9, 0], [9, 15, 0], [10, 16, 1], [11, 18, 0], [12, 13, 0],
-            [13, 20, 0], [14, 22, 1], [15, 24, 1], [16, 17, 1], [17, 25, 2], [18, 19, 0],
-            [19, 26, 0], [20, 21, 0], [21, 32, 1], [21, 30, 0], [22, 23, 2], [23, -1, 2],
-            [24, 34, 1], [25, -1, 2], [26, 27, 1], [27, 36, 1], [28, 29, 0], [29, 38, 0],
-            [29, 40, 0], [31, 44, 0], [31, 42, 0], [32, 33, 2], [33, -1, 2], [34, 46, 2],
-            [35, -1, 2], [35, 36, 2], [37, 48, 1], [37, 38, 1], [39, 50, 1], [39, 52, 1],
-            [41, 52, 0], [41, 54, 0], [43, 44, 0], [44, 45, 1], [45, -1, 1], [46, -1, 2],
-            [47, -1, 2], [47, 48, 2], [49, -1, 2], [49, 50, 2], [51, 60, 1], [51, 62, 0],
-            [53, 62, 0], [53, 64, 0], [54, 55, 0], [56, -1, 2], [56, 65, 1], [57, -1, 2],
-            [57, 67, 2], [58, -1, 2], [58, 59, 2], [59, 60, 1], [61, 69, 0], [61, 70, 0],
-            [63, 70, 0], [65, 66, 1], [66, 71, 0], [66, 73, 0], [67, 68, 1], [68, 69, 1],
-            [71, 72, 0]]
-    cables:
-        cable_type: [0, 1, 2]
-        cross_section: [95, 240, 500]
-        current_capacity: [300, 480, 655]
-        turbines_supplied: [3, 5, 7]
-        
-turbines: !include IEA37_10MW_turbine.yaml
