+++ examples/01_onshore/inputs/power_thrust_table_ccblade_IEA-3p4-130-RWT.csv
