+++ examples/01_onshore/inputs/turbine_spec_IEA-3p4-130-RWT.yaml
