+++ examples/01_onshore/inputs/wrg_example.wrg
