+++ examples/02_offshore_fixed/inputs/power_thrust_table_ccblade_IEA-22-284-RWT.csv
