+++ examples/02_offshore_fixed/inputs/turbine_spec_IEA-22-284-RWT.yaml
