+++ examples/02_offshore_fixed/inputs/wrg_example.wrg
