+++ examples/02_offshore_fixed/optimization_demo.py
@@ -17,23 +17,23 @@ def run_example():
     # set up system
     prob = set_up_ard_model(input_dict=input_dict, root_data_path="inputs")
 
+    # run the model
+    prob.run_model()
+
     if False:
         # visualize model
         om.n2(prob)
-
-    # run the model
-    prob.run_model()
 
     # collapse the test result data
     test_data = {
         "AEP_val": float(prob.get_val("AEP_farm", units="GW*h")[0]),
         "CapEx_val": float(prob.get_val("tcc.tcc", units="MUSD")[0]),
-        "BOS_val": float(prob.get_val("landbosse.total_capex", units="MUSD")[0]),
+        "BOS_val": float(prob.get_val("orbit.total_capex", units="MUSD")[0]),
         "OpEx_val": float(prob.get_val("opex.opex", units="MUSD/yr")[0]),
         "LCOE_val": float(prob.get_val("financese.lcoe", units="USD/MW/h")[0]),
         "area_tight": float(prob.get_val("landuse.area_tight", units="km**2")[0]),
         "coll_length": float(
-            prob.get_val("optiwindnet_coll.total_length_cables", units="km")[0]
+            prob.get_val("collection.total_length_cables", units="km")[0]
         ),
         "turbine_spacing": float(
             np.min(prob.get_val("spacing_constraint.turbine_spacing", units="km"))
@@ -56,12 +56,12 @@ def run_example():
         test_data = {
             "AEP_val": float(prob.get_val("AEP_farm", units="GW*h")[0]),
             "CapEx_val": float(prob.get_val("tcc.tcc", units="MUSD")[0]),
-            "BOS_val": float(prob.get_val("landbosse.total_capex", units="MUSD")[0]),
+            "BOS_val": float(prob.get_val("orbit.total_capex", units="MUSD")[0]),
             "OpEx_val": float(prob.get_val("opex.opex", units="MUSD/yr")[0]),
             "LCOE_val": float(prob.get_val("financese.lcoe", units="USD/MW/h")[0]),
             "area_tight": float(prob.get_val("landuse.area_tight", units="km**2")[0]),
             "coll_length": float(
-                prob.get_val("optiwindnet_coll.total_length_cables", units="km")[0]
+                prob.get_val("collection.total_length_cables", units="km")[0]
             ),
             "turbine_spacing": float(
                 np.min(prob.get_val("spacing_constraint.turbine_spacing", units="km"))
@@ -76,43 +76,43 @@ def run_example():
         pp.pprint(test_data)
         print("\n\n")
 
+        # plot convergence
+        ## read cases
+        cr = om.CaseReader(
+            prob.get_outputs_dir()
+            / input_dict["analysis_options"]["recorder"]["filepath"]
+        )
 
-#         # plot convergence
-#         ## read cases
-#         cr = om.CaseReader(
-#             prob.get_outputs_dir()
-#             / input_dict["analysis_options"]["recorder"]["filepath"]
-#         )
-#
-#         # Extract the driver cases
-#         cases = cr.get_cases("driver")
-#
-#         # Initialize lists to store iteration data
-#         iterations = []
-#         objective_values = []
-#
-#         # Loop through the cases and extract iteration number and objective value
-#         for i, case in enumerate(cases):
-#             iterations.append(i)
-#             objective_values.append(
-#                 case.get_objectives()[
-#                     input_dict["analysis_options"]["objective"]["name"]
-#                 ]
-#             )
-#
-#         # Plot the convergence
-#         plt.figure(figsize=(8, 6))
-#         plt.plot(iterations, objective_values, marker="o", label="Objective (LCOE)")
-#         plt.xlabel("Iteration")
-#         plt.ylabel("Objective Value (Total Cable Length (m))")
-#         plt.title("Convergence Plot")
-#         plt.legend()
-#         plt.grid()
-#         plt.show()
-#
-#     optiwindnet.plotting.gplot(prob.model.optiwindnet_coll.graph)
-#
-#     plt.show()
+        # Extract the driver cases
+        cases = cr.get_cases("driver")
+
+        # Initialize lists to store iteration data
+        iterations = []
+        objective_values = []
+
+        # Loop through the cases and extract iteration number and objective value
+        for i, case in enumerate(cases):
+            iterations.append(i)
+            objective_values.append(
+                case.get_objectives()[
+                    input_dict["analysis_options"]["objective"]["name"]
+                ]
+            )
+
+        # Plot the convergence
+        plt.figure(figsize=(8, 6))
+        plt.plot(iterations, objective_values, marker="o", label="Objective (LCOE)")
+        plt.xlabel("Iteration")
+        plt.ylabel("Objective Value (Total Cable Length (m))")
+        plt.title("Convergence Plot")
+        plt.legend()
+        plt.grid()
+        plt.show()
+
+    optiwindnet.plotting.gplot(prob.model.collection.graph)
+
+    plt.show()
+
 
 if __name__ == "__main__":
