+++ examples/03_offshore_floating/inputs/power_thrust_table_ccblade_IEA-22-284-RWT.csv
