+++ examples/03_offshore_floating/inputs/turbine_spec_IEA-22-284-RWT.yaml
