+++ examples/03_offshore_floating/inputs/wrg_example.wrg
