+++ examples/03_offshore_floating/optimization_demo.py
@@ -3,10 +3,11 @@ import pprint as pp
 import numpy as np
 import matplotlib.pyplot as plt
 
+import openmdao.api as om
+
 import optiwindnet.plotting
 from ard.utils.io import load_yaml
 from ard.api import set_up_ard_model
-import openmdao.api as om
 
 
 def run_example():
@@ -15,13 +16,17 @@ def run_example():
     input_dict = load_yaml("./inputs/ard_system.yaml")
 
     # set up system
-    prob = set_up_ard_model(input_dict=input_dict, root_data_path="inputs")
+    prob = set_up_ard_model(
+        input_dict=input_dict,
+        root_data_path="inputs",
+    )
 
     # run the model
     prob.run_model()
 
-    # Visualize model
-    om.n2(prob)
+    if False:
+        # visualize model
+        om.n2(prob)
 
     # collapse the test result data
     test_data = {
@@ -33,6 +38,9 @@ def run_example():
         "area_tight": float(prob.get_val("landuse.area_tight", units="km**2")[0]),
         "coll_length": float(
             prob.get_val("collection.total_length_cables", units="km")[0]
+        ),
+        "mooring_spacing": float(
+            np.min(prob.get_val("mooring_constraint.mooring_spacing", units="km"))
         ),
         "turbine_spacing": float(
             np.min(prob.get_val("spacing_constraint.turbine_spacing", units="km"))
@@ -51,8 +59,14 @@ def run_example():
         prob.run_driver()
         prob.cleanup()
 
+        prob.check_totals(compact_print=True, show_only_incorrect=True)
+
         # collapse the test result data
         test_data = {
+            "spacing_primary": float(prob.get_val("spacing_primary")[0]),
+            "spacing_secondary": float(prob.get_val("spacing_secondary")[0]),
+            "angle_orientation": float(prob.get_val("angle_orientation")[0]),
+            "angle_skew": float(prob.get_val("angle_skew")[0]),
             "AEP_val": float(prob.get_val("AEP_farm", units="GW*h")[0]),
             "CapEx_val": float(prob.get_val("tcc.tcc", units="MUSD")[0]),
             "BOS_val": float(prob.get_val("orbit.total_capex", units="MUSD")[0]),
@@ -61,6 +75,9 @@ def run_example():
             "area_tight": float(prob.get_val("landuse.area_tight", units="km**2")[0]),
             "coll_length": float(
                 prob.get_val("collection.total_length_cables", units="km")[0]
+            ),
+            "mooring_spacing": float(
+                np.min(prob.get_val("mooring_constraint.mooring_spacing", units="km"))
             ),
             "turbine_spacing": float(
                 np.min(prob.get_val("spacing_constraint.turbine_spacing", units="km"))
@@ -116,61 +133,3 @@ def run_example():
 if __name__ == "__main__":
 
     run_example()
-
-# RESULTS:
-
-
-# {'AEP_val': 4818.0,
-#  'BOS_val': 2127.5924853696597,
-#  'CapEx_val': 768.4437570425,
-#  'LCOE_val': 57.63858824842508,
-#  'OpEx_val': 60.50000000000001,
-#  'area_tight': 63.234304,
-#  'coll_length': 47.761107521256534,
-#  'mooring_spacing': 1.1208759839268934}
-
-
-# /opt/anaconda3/envs/ard/lib/python3.12/site-packages/openmdao/recorders/sqlite_recorder.py:231: UserWarning:The existing case recorder file, /Users/jthomas2/Documents/programs/Ard/examples/offshore/optimization_demo_out/opt_results.sql, is being overwritten.
-# Optimization terminated successfully    (Exit mode 0)
-#             Current function value: 20495.94468696311
-#             Iterations: 13
-#             Function evaluations: 9
-#             Gradient evaluations: 9
-# Optimization Complete
-# -----------------------------------
-
-
-# RESULTS (opt):
-
-# {'AEP_val': 4818.0,
-#  'BOS_val': 2113.9686662769814,
-#  'CapEx_val': 768.4437570425,
-#  'LCOE_val': 57.42651136342073,
-#  'OpEx_val': 60.50000000000001,
-#  'area_tight': 11.614464,
-#  'coll_length': 20.49594468696311,
-#  'mooring_spacing': 0.0582385263254008,
-#  'turbine_spacing': 0.8519999999999998}
-
-
-## 20250714
-# RESULTS:
-
-# {'AEP_val': 4818.0,
-#  'BOS_val': 1431.448205129097,
-#  'CapEx_val': 768.4437570425,
-#  'LCOE_val': 46.80197118365915,
-#  'OpEx_val': 60.50000000000001,
-#  'area_tight': 63.234304,
-#  'coll_length': 47.712041428901635}
-
-# RESULTS (opt):
-
-# {'AEP_val': 4818.0,
-#  'BOS_val': 1412.8143111685533,
-#  'CapEx_val': 768.4437570425,
-#  'LCOE_val': 46.51190434118493,
-#  'OpEx_val': 60.50000000000001,
-#  'area_tight': 11.614464,
-#  'coll_length': 20.507898644867613,
-#  'turbine_spacing': 0.8519999999999999}
