+++ pyproject.toml
@@ -13,12 +13,16 @@ build-backend = "setuptools.build_meta"
 where = ["."]
 include = ["ard", "ard.*"]
 
+[tool.setuptools.package-data]
+"ard.api.default_systems" = ["*.yaml"]
+
 [project]
 name = "Ard"
 version = "1.0.0-alpha0"
 authors = [
   {name = "Cory Frontin", email = "cory.frontin@nrel.gov"},
   {name = "Rafael Mudafort", email = "rafael.mudafort@nrel.gov"},
+  {name = "Jared Thomas", email = "jared.thomas@nrel.gov"},
 ]
 description = "A package for multidisciplinary and/or multifidelity wind farm design"
 readme= "README.md"
@@ -43,7 +47,7 @@ dependencies = [
   "shapely",
   "jax",
   "optiwindnet>=0.0.4",
-  "statsmodels@git+https://github.com/statsmodels/statsmodels.git@main",
+  "statsmodels",
   "highspy",
   "pyyaml",
 ]
